Move to internal/cmd package
