+++ cmd/execute.go
@@ -5,12 +5,12 @@ import (
 	"context"
 
 	"go.k6.io/k6/cmd/state"
-	"go.k6.io/k6/internal/launcher"
+	internalcmd "go.k6.io/k6/internal/cmd"
 )
 
 // Execute executes the k6 command
 // It only is exported here for backwards compatibility and the ability to use xk6 to build extended k6
 func Execute() {
 	gs := state.NewGlobalState(context.Background())
-	launcher.New(gs).Launch()
+	internalcmd.NewLauncher(gs).Launch()
 }
