+++ internal/cmd/analyze_test.go
@@ -1,4 +1,4 @@
-package launcher
+package cmd
 
 import (
 	"testing"
